pyproject.toml
[tool.poetry]
name = "xarxes2025"
version = "0.1.0"
description = "Esquelet per xarxes 2025 @ UdL.cat"
authors = ["Carles Mateu <carles.mateu@udl.cat>"]
readme = "README.md"
packages = [
    {include = "xarxes2025", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.12"
click = ">=8.1.8,<9.0.0"
loguru = ">=0.7.3,<0.8.0"
pillow = ">=11.1.0,<12.0.0"
opencv-python = ">=4.11.0.86,<5.0.0.0"

[tool.poetry.group.dev.dependencies]
icecream = "^2.1.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
xarxes2025 = "xarxes2025.cli:cli"